YQL-20279 exclude internal YQL protos for java, moved issue_id.proto to public
в основном они протекают после включения contrib/ydb/core/protos (что на самом деле не публичная часть GRPC сервиса)
commit_hash:1fec0018e2f898417521c5b0b39d6ba64fb97c80
