--- yql/essentials/core/issue/protos/ya.make
@@ -1,13 +1,0 @@
-PROTO_LIBRARY()
-
-SRCS(
-    issue_id.proto
-)
-
-PEERDIR(
-    yql/essentials/public/issue/protos
-)
-
-EXCLUDE_TAGS(GO_PROTO)
-
-END()
