+++ yql/essentials/core/issue/yql_issue.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <yql/essentials/core/issue/protos/issue_id.pb.h>
+#include <yql/essentials/public/issue/protos/issue_id.pb.h>
 #include <yql/essentials/public/issue/yql_issue.h>
 #include <yql/essentials/public/issue/yql_issue_id.h>
