+++ yql/essentials/core/type_ann/ya.make
@@ -27,7 +27,7 @@ PEERDIR(
     yql/essentials/core
     yql/essentials/core/expr_nodes
     yql/essentials/core/issue
-    yql/essentials/core/issue/protos
+    yql/essentials/public/issue/protos
     yql/essentials/core/sql_types
     yql/essentials/core/poly_args
     yql/essentials/providers/common/schema/expr
