+++ yql/essentials/minikql/jsonpath/executor.cpp
@@ -1,7 +1,7 @@
 #include "executor.h"
 #include <yql/essentials/minikql/jsonpath/parser/parse_double.h>
 
-#include <yql/essentials/core/issue/protos/issue_id.pb.h>
+#include <yql/essentials/public/issue/protos/issue_id.pb.h>
 #include <yql/essentials/minikql/dom/node.h>
 
 #include <util/generic/scope.h>
