+++ yql/essentials/minikql/jsonpath/parser/type_check.cpp
@@ -1,6 +1,6 @@
 #include "type_check.h"
 
-#include <yql/essentials/core/issue/protos/issue_id.pb.h>
+#include <yql/essentials/public/issue/protos/issue_id.pb.h>
 
 namespace NYql::NJsonPath {
