+++ yql/essentials/minikql/jsonpath/parser/ya.make
@@ -19,7 +19,7 @@ PEERDIR(
     yql/essentials/public/issue
     yql/essentials/ast
     yql/essentials/utils
-    yql/essentials/core/issue/protos
+    yql/essentials/public/issue/protos
     yql/essentials/parser/proto_ast/antlr3
     yql/essentials/parser/proto_ast/gen/jsonpath
 )
