+++ yql/essentials/public/issue/protos/ya.make
@@ -1,6 +1,7 @@
 PROTO_LIBRARY()
 
 SRCS(
+    issue_id.proto
     issue_message.proto
     issue_severity.proto
 )
