+++ yql/essentials/sql/settings/ya.make
@@ -15,7 +15,7 @@ PEERDIR(
     yql/essentials/public/udf_meta
     yql/essentials/core/issue
     yql/essentials/core/pg_settings
-    yql/essentials/core/issue/protos
+    yql/essentials/public/issue/protos
     yql/essentials/utils
 )
