+++ yql/essentials/sql/v0/lexer/ya.make
@@ -4,7 +4,7 @@ ENABLE(SKIP_YQL_STYLE_CPP)
 NO_CLANG_TIDY()
 
 PEERDIR(
-    yql/essentials/core/issue/protos
+    yql/essentials/public/issue/protos
     yql/essentials/parser/proto_ast/gen/v0
 )
