+++ yql/essentials/types/binary_json/ut_benchmark/ya.make
@@ -21,7 +21,7 @@ PEERDIR(
     yql/essentials/minikql/dom
     yql/essentials/minikql/invoke_builtins/llvm16
     yql/essentials/public/udf/service/exception_policy
-    yql/essentials/core/issue/protos
+    yql/essentials/public/issue/protos
     yql/essentials/sql/pg_dummy
 )
